requirements.txt
databricks-sdk>=0.20.0
simple-salesforce>=1.12.0
slack-sdk>=3.27.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
requests>=2.31.0
pyspark>=3.5.0
